Implement --extra-package=package.deb
This simply adds a deb on the host system to the existing dummy archive
created for the build-dependency resolution metapackage, which makes it
available for resolving build dependencies. The purpose of this option
is to make it easier to build packages against locally-built build
depenencies, without waiting for those packages to enter the main
archive, or going through the hassle of maintaining a local archive and
making it accessible inside the chroot.

This option is named similarly to the EXTRAPACKAGES option in
pbuilderrc(5), but note that it takes a .deb, not a package name.

Thanks to Emanuele Aina for review!

Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>