Implement --extra-repository="deb scheme://..."
This simply adds another line to the sources.list.d file that lists our
dummy archive for build-dependency resolution, making that repository
available for providing build-dependencies. The purpose of this option
is to make it easier to build packages against a local archive without
modifying the source chroot to list that archive.

Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>